+++ wgpu-core/src/command/compute.rs
@@ -308,6 +308,14 @@ impl Global {
         };
 
         arc_desc.timestamp_writes = if let Some(tw) = desc.timestamp_writes {
+            match cmd_buf
+                .device
+                .require_features(wgt::Features::TIMESTAMP_QUERY)
+            {
+                Ok(()) => (),
+                Err(e) => return make_err(e.into(), arc_desc),
+            }
+
             let query_set = match hub.query_sets.get(tw.query_set).get() {
                 Ok(query_set) => query_set,
                 Err(e) => return make_err(e.into(), arc_desc),